homework3.1

+++ Lection3/Homework 3.1.py
@@ -2,11 +2,11 @@ num_1=int(input('number1'))
 effect =input('effect')
 num_2=int(input('number2'))
 if effect == '+':
-   print(num_1 + num_2)
+   print(int(num_1 + num_2))
 elif effect == '-':
-    print(num_1-num_2)
+    print(int(num_1-num_2))
 elif effect=='*':
-    print(num_1*num_2)
+    print(int(num_1*num_2))
 elif effect=='/':
-    print(num_1/num_2)
+    print(int(num_1/num_2))
